refactor(adapters): remove `DomainError`, streamline error handling with `String`, delete unused services (`news`, `source`), and update references across repositories and services
