+++ crates/adapters/src/services.rs
@@ -1,4 +1,2 @@
 pub mod http;
-pub mod news;
-pub mod source;
 pub mod storage;
