+++ crates/adapters/src/utils/sqlite.rs
@@ -1,0 +1,2 @@
+pub mod codecs;
+pub mod tools;
