+++ crates/adapters/src/utils/sqlite/codecs.rs
