--- crates/core/src/errors.rs
@@ -1,22 +1,0 @@
-use std::fmt;
-
-#[derive(Debug)]
-pub enum DomainError {
-    NotFound(String),
-    InvalidInput(String),
-    Conflict(String),
-    Internal(String),
-}
-
-impl fmt::Display for DomainError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Self::NotFound(msg) => write!(f, "Not found: {msg}"),
-            Self::InvalidInput(msg) => write!(f, "Invalid input: {msg}"),
-            Self::Conflict(msg) => write!(f, "Conflict: {msg}"),
-            Self::Internal(msg) => write!(f, "Internal error: {msg}"),
-        }
-    }
-}
-
-impl std::error::Error for DomainError {}
