+++ crates/core/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod errors;
 pub mod models;
 pub mod repos;
 pub mod services;
