+++ crates/core/src/repos/source.rs
@@ -1,18 +1,13 @@
 use super::SourceCreate;
-use crate::errors::DomainError;
 use crate::models::source::{SourceType, SourceTypeValue};
 use async_trait::async_trait;
 use uuid::Uuid;
 
 #[async_trait]
 pub trait SourceRepository: Send + Sync {
-    async fn add(&self, draft: SourceCreate) -> Result<(), DomainError>;
-    async fn get_by_id(&self, id: Uuid) -> Result<SourceType, DomainError>;
-    async fn get_all(&self) -> Result<Vec<SourceType>, DomainError>;
-    async fn delete_by_id(&self, id: Uuid) -> Result<(), DomainError>;
-    async fn delete_with_type(
-        &self,
-        id: Uuid,
-        source_type: SourceTypeValue,
-    ) -> Result<(), DomainError>;
+    async fn add(&self, draft: SourceCreate) -> Result<(), String>;
+    async fn get_by_id(&self, id: Uuid) -> Result<SourceType, String>;
+    async fn get_all(&self) -> Result<Vec<SourceType>, String>;
+    async fn delete_by_id(&self, id: Uuid) -> Result<(), String>;
+    async fn delete_with_type(&self, id: Uuid, source_type: SourceTypeValue) -> Result<(), String>;
 }
