+++ crates/core/src/services.rs
@@ -1,7 +1,3 @@
-use crate::models::News;
-use crate::models::source::atom::AtomDraft;
-use crate::models::source::telegram::TelegramChannelDraft;
-use crate::models::source::{SourceType, SourceTypeValue};
 use async_trait::async_trait;
 use reqwest::Response;
 use std::sync::Arc;
@@ -25,25 +21,4 @@ pub trait HttpService: Send + Sync {
 #[async_trait]
 pub trait StorageService: Send + Sync {
     async fn save_raw(&self, key: Uuid, value: &str);
-}
-
-#[async_trait]
-pub trait NewsService: Send + Sync {
-    async fn update_news(&self, news: &[Arc<dyn News>]) -> Result<(), String>;
-}
-
-pub enum SourceCreate {
-    Atom(AtomDraft),
-    TelegramChannel(TelegramChannelDraft),
-}
-
-#[async_trait]
-pub trait SourceService: Send + Sync {
-    async fn add(&self, draft: SourceCreate) -> Result<(), String>;
-
-    async fn get_by_id(&self, id: Uuid) -> Result<SourceType, String>;
-    async fn get_all(&self) -> Result<Vec<SourceType>, String>;
-
-    async fn delete_by_id(&self, id: Uuid) -> Result<(), String>;
-    async fn delete_with_type(&self, id: Uuid, source_type: SourceTypeValue) -> Result<(), String>;
 }
